# PERF: PeriodIndex.dropna, difference, take