--- a/pandas/util/testing.py
+++ b/pandas/util/testing.py
@@ -824,7 +824,8 @@ def assert_index_equal(left, right, exact='equiv', check_names=True,
         # accept level number only
         unique = index.levels[level]
         labels = index.labels[level]
-        filled = take_1d(unique.values, labels, fill_value=unique._na_value)
+        filled = take_1d(unique._take_values, labels,
+                         fill_value=unique._na_value)
         values = unique._shallow_copy(filled, name=index.names[level])
         return values