@pep8speaks 2018-10-11
Hello @jbrockmendel! Thanks for submitting the PR.

 - There are no PEP8 issues in the file [`asv_bench/benchmarks/period.py`](https://github.com/jbrockmendel/pandas/blob/f3d332142b3272bb5764168f862cafdfdfead110/asv_bench/benchmarks/period.py) !

 - There are no PEP8 issues in the file [`pandas/core/indexes/base.py`](https://github.com/jbrockmendel/pandas/blob/f3d332142b3272bb5764168f862cafdfdfead110/pandas/core/indexes/base.py) !

 - There are no PEP8 issues in the file [`pandas/core/indexes/multi.py`](https://github.com/jbrockmendel/pandas/blob/f3d332142b3272bb5764168f862cafdfdfead110/pandas/core/indexes/multi.py) !

 - There are no PEP8 issues in the file [`pandas/core/indexes/period.py`](https://github.com/jbrockmendel/pandas/blob/f3d332142b3272bb5764168f862cafdfdfead110/pandas/core/indexes/period.py) !

 - There are no PEP8 issues in the file [`pandas/util/testing.py`](https://github.com/jbrockmendel/pandas/blob/f3d332142b3272bb5764168f862cafdfdfead110/pandas/util/testing.py) !

 - Complete extra results for this file :

> file_to_check.py:2454:-258: W605 invalid escape sequence '\('
file_to_check.py:2454:-255: W605 invalid escape sequence '\)'
---