@jorisvandenbossche 2018-10-11
And the same question as on the other PR: to what extent is this conflicting with the PeriodArray PR of @TomAugspurger ? Shouldn't we rather focus on finishing that one first, and then doing some perf improvements / clean-ups afterwards?