@jbrockmendel 2018-10-11
> Shouldn't we rather focus on finishing that one first, and then doing some perf improvements

@jorisvandenbossche @TomAugspurger responding to many of joris's comments/questions (here and in #23093) in one place.

The PeriodIndex/PeriodArray constructors are a pain point in large part because its `_simple_new` is not very simple, which is why there is the lower-level `_from_ordinals`.  Why the complication in `_simple_new`?  Because the values being passed to it are sometimes i8 (what we actually want), sometimes `ndarray[object]`, and sometimes `PeriodIndex`.  The latter two cases, as it turns out, are coming from calls to `_shallow_copy`, exactly the ones this PR affects.

i.e. following this PR, we'll be able to clean up the `PeriodIndex` constructors (on top of whats in #23093)