@jorisvandenbossche 2018-10-11
@jbrockmendel to be clear, I fully agree with your analysis about `_simple_new`. 
But the thing is that the PeriodArray PR already solves this partly (at least for PeriodIndex, complexity is mainly moved to PeriodArray), because then, it will be much more PeriodArray that will be passed around in the Index implementation, instead of sometimes object array / sometimes ordinals. And of course, we can then still try to simplify the different constructors in PeriodArray. 
As another example, The PeriodArray currently does not yet solve the `take` issue you try to solve here, but once that PR would land, the solution for `take` would be completely different (implementing it on the PeriodArray, dispatching to that, and properly wrapping it again), and the other changes you have here might not be needed.