@jbrockmendel 2018-10-11
> And of course, we can then still try to simplify the different constructors in PeriodArray.

That's fair.  I am definitely viewing the situation as "#22862 is a big PR and simplifying things before that will make that PR smaller, review easier, etc"