@jorisvandenbossche 2018-10-12
Sorry, I still don't see how introducing this `_take_values` brings us closer to our goal. The PeriodArray PR will remove that again, no?