@TomAugspurger 2018-10-12
I think this idea is that they'll all just be `.values` once we have
PeriodArray / DatetimeArray.

On Fri, Oct 12, 2018 at 7:50 AM Jeff Reback <notifications@github.com>
wrote:

> *@jreback* requested changes on this pull request.
>
> I don't agree this will be short lived. This is a new notion, meaning we
> now have
> ------------------------------
>
> In pandas/core/indexes/base.py
> <https://github.com/pandas-dev/pandas/pull/23095#discussion_r224771799>:
>
> > @@ -763,6 +763,17 @@ def get_values(self):
>          """
>          return self.values
>
> +    @property
> +    def _take_values(self):
> +        """
> +        Values to use in `take` operation, suitable to being passed back to
> +        _shallow_copy/_simple_new.
> +        """
> +        if is_period_dtype(self):
> +            # Avoid casting to object-dtype
> +            return self._ndarray_values
> +        return self.values
>
> this is basically .asi8 on Indexes now, yes? Is there a reason we are not
> aligning on that? or changing .asi8 to this?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/23095#pullrequestreview-164225426>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABQHIgYuFBTeDDuieFgaOcWO-fKYWR-cks5ukJAegaJpZM4XYCM1>
> .
>