@jbrockmendel 2018-10-12
> I don't agree this will be short lived. 

> I think this idea is that they'll all just be `.values` once we have PeriodArray / DatetimeArray.

I tend to agree with Tom on this one, but the more important point in my book is that this fixes the weird arguments being passed to PeriodIndex._shallow_copy which allows us to clean up the constructor mess, simplifying the PeriodArray transition.