@jreback 2018-10-14
if they will be .values then don’t introduce something new