@jbrockmendel 2018-10-14
Yah, I'm beginning to agree.  Plenty of other related stuff to focus on.  Closing.